Revert "Revert "Move most `Verb`s into one utility class.""
This reverts commit b592c3c080097390d734aa3c78c9ecb358a5247e.
